Display Thai ID card after confirmation
