+++ frontend_vite/src/App.jsx
@@ -1,56 +1,9 @@
 import { useEffect, useState } from 'react'
 import cardImage from './assets/react.svg'
+import ThaiIDCard from './components/ThaiIDCard'
 import './App.css'
 
 export default function App() {
-  const getField = (obj, keys) => {
-    for (const key of keys) {
-      if (obj && obj[key]) return obj[key]
-    }
-    return ''
-  }
-
-  const formatCid = (cid) => {
-    if (!cid) return ''
-    const digits = cid.toString().replace(/\D/g, '')
-    if (digits.length !== 13) return cid
-    return digits.replace(/(\d{1})(\d{4})(\d{5})(\d{2})(\d{1})/, '$1 $2 $3 $4 $5')
-  }
-
-  const formatDob = (dob) => {
-    if (!dob) return ''
-    const str = dob.toString()
-    if (str.length !== 8) return dob
-    let year = parseInt(str.slice(0, 4), 10)
-    const month = parseInt(str.slice(4, 6), 10) - 1
-    const day = parseInt(str.slice(6, 8), 10)
-    if (year > 2400) year -= 543
-    const date = new Date(year, month, day)
-    return date.toLocaleDateString('th-TH', {
-      day: 'numeric',
-      month: 'long',
-      year: 'numeric',
-    })
-  }
-
-  const formatAddress = (info) => {
-    if (!info) return ''
-    if (info.address)
-      return info.address.replace(/undefined/g, '').replace(/\s+/g, ' ').trim()
-    const houseNo = getField(info, ['houseNo', 'hno'])
-    const villageNo = getField(info, ['moo', 'villageNo'])
-    const subDistrict = getField(info, ['tambon', 'subdistrict'])
-    const district = getField(info, ['amphur', 'district'])
-    const province = getField(info, ['province'])
-    const parts = []
-    if (houseNo) parts.push(houseNo)
-    if (villageNo) parts.push(`หมู่ที่ ${villageNo}`)
-    if (subDistrict) parts.push(`ตำบล${subDistrict}`)
-    if (district) parts.push(`อำเภอ${district}`)
-    if (province) parts.push(`จังหวัด${province}`)
-    return parts.join(' ')
-  }
-
   const [now, setNow] = useState(new Date())
   const [hospital, setHospital] = useState(null)
   const [dbError, setDbError] = useState(false)
@@ -131,30 +84,7 @@ export default function App() {
           <button className="btn secondary">พิมพ์บัตรคิว</button>
           <button className="btn muted">ปิดสิทธิ(ยืนยันตัวตน)</button>
         </div>
-        {cardInfo && (
-          <div className="card-info">
-            <div className="row">
-              <span className="label">Identification Number</span>
-              <span>{formatCid(getField(cardInfo, ['cid', 'pid', 'nationalId']))}</span>
-            </div>
-            <div className="row">
-              <span className="label">Name</span>
-              <span>{getField(cardInfo, ['firstname', 'fname', 'firstNameTH', 'name'])}</span>
-            </div>
-            <div className="row">
-              <span className="label">Last name</span>
-              <span>{getField(cardInfo, ['lastname', 'lname', 'lastNameTH', 'surname'])}</span>
-            </div>
-            <div className="row">
-              <span className="label">Date of Birth</span>
-              <span>{formatDob(getField(cardInfo, ['birthdate', 'birthDate', 'dob']))}</span>
-            </div>
-            <div className="row">
-              <span className="label">Address</span>
-              <span>{formatAddress(cardInfo)}</span>
-            </div>
-          </div>
-        )}
+        {cardInfo && <ThaiIDCard info={cardInfo} />}
       </main>
     </div>
   )
